GCP Lambda-Deteccion de fraudes/requirements.txt
apache-beam[gcp]>=2.46.0
google-cloud-pubsub
google-cloud-bigquery
